2515 moved logic to MatchOperator from MatchOperation
